chore: rearrange files
