+++ satellite/src/client/parser.ts
@@ -1,0 +1,59 @@
+export function parseLineParameters(line: string): Record<string, string | boolean> {
+	const makeSafe = (index: number): number => {
+		return index === -1 ? Number.POSITIVE_INFINITY : index
+	}
+
+	const fragments: string[] = ['']
+	let quotes = 0
+
+	let i = 0
+	while (i < line.length) {
+		// Find the next characters of interest
+		const spaceIndex = makeSafe(line.indexOf(' ', i))
+		const slashIndex = makeSafe(line.indexOf('\\', i))
+		const quoteIndex = makeSafe(line.indexOf('"', i))
+
+		// Find which is closest
+		const o = Math.min(spaceIndex, slashIndex, quoteIndex)
+		if (!isFinite(o)) {
+			// None were found, copy the remainder and stop
+			const slice = line.substring(i)
+			fragments[fragments.length - 1] += slice
+
+			break
+		} else {
+			// copy the slice before this character
+			const slice = line.substring(i, o)
+			fragments[fragments.length - 1] += slice
+
+			const c = line[o]
+			if (c == '\\') {
+				// If char is a slash, the character following it is of interest
+				// Future: does this consider non \" chars?
+				fragments[fragments.length - 1] += line[o + 1]
+
+				i = o + 2
+			} else {
+				i = o + 1
+
+				// Figure out what the char was
+				if (c === '"') {
+					quotes ^= 1
+				} else if (!quotes && c === ' ') {
+					fragments.push('')
+				} else {
+					fragments[fragments.length - 1] += c
+				}
+			}
+		}
+	}
+
+	const res: Record<string, string | boolean> = {}
+
+	for (const fragment of fragments) {
+		const [key, value] = fragment.split('=', 2)
+		res[key] = value === undefined ? true : value
+	}
+
+	return res
+}
