+++ satellite/src/config.ts
@@ -2,7 +2,7 @@ import Conf, { Schema } from 'conf'
 import path from 'path'
 import os from 'os'
 import { customAlphabet } from 'nanoid'
-import { SomeConnectionDetails } from './clientImplementations.js'
+import { SomeConnectionDetails } from './client/socketImplementations.js'
 import { assertNever, DEFAULT_TCP_PORT, DEFAULT_WS_PORT } from './lib.js'
 import debounceFn from 'debounce-fn'
 import { setMaxListeners } from 'events'
